REST API: User By id, all cards, cards by id Persistence for Cards Service for Cards and User

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -32,8 +32,8 @@ CREATE TABLE cards (
 
 CREATE TABLE friend_request (
           id BIGINT NOT NULL DEFAULT nextval('friend_request_id_seq'::regclass) PRIMARY key,
-          user_1_ref int NOT NULL,
-          user_2_ref int NOT NULL,
+          user_1_ref BIGINT NOT NULL,
+          user_2_ref BIGINT NOT NULL,
           pending int NOT NULL,
           CONSTRAINT friends_user_1_ref FOREIGN KEY (user_1_ref) REFERENCES users (id),
           CONSTRAINT friends_user_2_ref FOREIGN KEY (user_2_ref) REFERENCES users (id)
@@ -42,24 +42,24 @@ CREATE TABLE friend_request (
 CREATE TABLE sessions (
           id BIGINT NOT NULL DEFAULT nextval('sessions_id_seq'::regclass) PRIMARY key,
           session_name varchar(100) NOT NULL,
-          user_1_id int DEFAULT NULL,
-          user_2_id int DEFAULT NULL,
+          user_1_id BIGINT DEFAULT NULL,
+          user_2_id BIGINT DEFAULT NULL,
           CONSTRAINT sessions_user_1_ref FOREIGN KEY (user_1_id) REFERENCES users (id),
           CONSTRAINT sessions_user_2_ref FOREIGN KEY (user_2_id) REFERENCES users (id)
 );
 
 CREATE TABLE user_achievements (
           id BIGINT NOT NULL DEFAULT nextval('user_achievements_id_seq'::regclass) PRIMARY key,
-          user_ref int NOT NULL,
-          achievement_ref int NOT NULL,
+          user_ref BIGINT NOT NULL,
+          achievement_ref BIGINT NOT NULL,
           CONSTRAINT achievements_achievement_ref FOREIGN KEY (achievement_ref) REFERENCES achievements (id),
           CONSTRAINT achievements_user_ref FOREIGN KEY (user_ref) REFERENCES users (id)
 );
 
 CREATE TABLE user_cards (
           id BIGINT NOT NULL DEFAULT nextval('user_cards_id_seq'::regclass) PRIMARY key,
-          user_ref int NOT NULL,
-          card_ref int NOT NULL,
+          user_ref BIGINT NOT NULL,
+          card_ref BIGINT NOT NULL,
           CONSTRAINT cards_card_ref FOREIGN KEY (card_ref) REFERENCES cards (id),
           CONSTRAINT cards_user_ref FOREIGN KEY (user_ref) REFERENCES users (id)
 );
